Remove copy-pasted comments that literally repeat the code
These comments *literally* repeat what the code does.
Because of this the comments don't add information. They are
effectively noise. This is worse than not having these
comments.

A good argument why it's better to *not* have such comments
can be seen in this patch: some of these comments are
actually wrong. E.g. they mention the wrong class or
property (obviously not edited after copy-pasting). Or they
talk about "validation" that doesn't happen in the code.

Another argument is that implementation details should be
that: implementation details, and not be exposed in a
comment. I mean, who is supposed to read these comments?

Change-Id: Iad353c5d8b621ba5a2352540af4dd6cf99be80dd